requirements.txt
flask
discord-py-interactions